+++ src/os/file_windows.go
@@ -7,7 +7,6 @@ package os
 import (
 	"internal/poll"
 	"internal/syscall/windows"
-	"internal/testlog"
 	"runtime"
 	"syscall"
 	"unicode/utf16"
@@ -149,14 +148,8 @@ func openDir(name string) (file *File, err error) {
 	return f, nil
 }
 
-// OpenFile is the generalized open call; most users will use Open
-// or Create instead. It opens the named file with specified flag
-// (O_RDONLY etc.) and perm, (0666 etc.) if applicable. If successful,
-// methods on the returned File can be used for I/O.
-// If there is an error, it will be of type *PathError.
-func OpenFile(name string, flag int, perm FileMode) (*File, error) {
-	testlog.Open(name)
-
+// openFileNolog is the Windows implementation of OpenFile.
+func openFileNolog(name string, flag int, perm FileMode) (*File, error) {
 	if name == "" {
 		return nil, &PathError{"open", name, syscall.ENOENT}
 	}
