+++ src/os/getwd.go
@@ -30,13 +30,13 @@ func Getwd() (dir string, err error) {
 
 	// Clumsy but widespread kludge:
 	// if $PWD is set and matches ".", use it.
-	dot, err := Stat(".")
+	dot, err := statNolog(".")
 	if err != nil {
 		return "", err
 	}
 	dir = Getenv("PWD")
 	if len(dir) > 0 && dir[0] == '/' {
-		d, err := Stat(dir)
+		d, err := statNolog(dir)
 		if err == nil && SameFile(dot, d) {
 			return dir, nil
 		}
@@ -56,7 +56,7 @@ func Getwd() (dir string, err error) {
 	dir = getwdCache.dir
 	getwdCache.Unlock()
 	if len(dir) > 0 {
-		d, err := Stat(dir)
+		d, err := statNolog(dir)
 		if err == nil && SameFile(dot, d) {
 			return dir, nil
 		}
@@ -64,7 +64,7 @@ func Getwd() (dir string, err error) {
 
 	// Root is a special case because it has no parent
 	// and ends in a slash.
-	root, err := Stat("/")
+	root, err := statNolog("/")
 	if err != nil {
 		// Can't stat root - no hope.
 		return "", err
@@ -81,7 +81,7 @@ func Getwd() (dir string, err error) {
 		if len(parent) >= 1024 { // Sanity check
 			return "", syscall.ENAMETOOLONG
 		}
-		fd, err := Open(parent)
+		fd, err := openFileNolog(parent, O_RDONLY, 0)
 		if err != nil {
 			return "", err
 		}
@@ -93,7 +93,7 @@ func Getwd() (dir string, err error) {
 				return "", err
 			}
 			for _, name := range names {
-				d, _ := Lstat(parent + "/" + name)
+				d, _ := lstatNolog(parent + "/" + name)
 				if SameFile(d, dot) {
 					dir = "/" + name + dir
 					goto Found
