+++ src/os/stat_unix.go
@@ -7,7 +7,6 @@
 package os
 
 import (
-	"internal/testlog"
 	"syscall"
 )
 
@@ -26,10 +25,8 @@ func (f *File) Stat() (FileInfo, error) {
 	return &fs, nil
 }
 
-// Stat returns a FileInfo describing the named file.
-// If there is an error, it will be of type *PathError.
-func Stat(name string) (FileInfo, error) {
-	testlog.Stat(name)
+// statNolog stats a file with no test logging.
+func statNolog(name string) (FileInfo, error) {
 	var fs fileStat
 	err := syscall.Stat(name, &fs.sys)
 	if err != nil {
@@ -39,12 +36,8 @@ func Stat(name string) (FileInfo, error) {
 	return &fs, nil
 }
 
-// Lstat returns a FileInfo describing the named file.
-// If the file is a symbolic link, the returned FileInfo
-// describes the symbolic link. Lstat makes no attempt to follow the link.
-// If there is an error, it will be of type *PathError.
-func Lstat(name string) (FileInfo, error) {
-	testlog.Stat(name)
+// lstatNolog lstats a file with no test logging.
+func lstatNolog(name string) (FileInfo, error) {
 	var fs fileStat
 	err := syscall.Lstat(name, &fs.sys)
 	if err != nil {
